routes: Refactor /inner -> /components
